Fix the problem with trash can icon appearing on all the boards, when hovering one.
