+++ app/(auth)/layout.tsx
@@ -1,7 +1,7 @@
 'use client';
 
-import Image from 'next/image';
 import Link from 'next/link';
+import Image from 'next/image';
 
 const AuthLayout = ({ children }: { children: React.ReactNode }) => {
   return (
